python/requirements.txt
cryptography>=41.0.2
grpcio
sdc-apis
pyyaml
click
twine